Displacement maps, surface normal approximations
We introduce a displacement map upon the surface of the rendered
sphere through the new function calculate_displaced_radius_towards
(which takes a point and returns the displaced radius of the sphere
towards that point) and with a trivial modification to
signed_distance_from_sphere. Then, because the rendered object is no
longer a sphere, the unit surface normal at a hitpoint on the object's
surface is no longer just given by (hitpoint - sphere_center) /
sphere_radius. Instead, we introduce the function
get_unit_surface_normal, which uses the method of finite differences to
approximate the unit surface normal at a point on the sphere's surface
by computing the normalized approximated gradient of the signed distance
function at that point with respect to the three coordinate axes.
Finally, we remove the spatial texture that previously determined the
colors of points on the sphere's surface; now, we simply let Lambertian
reflectance determine the colors on our displacement-mapped sphere.