Enable disconnected objects from displacement maps
Previously, we restricted our displacement map's domain to the points on
the surface of the sphere. As our displacement map was continuous, this
guaranteed that the resulting displaced sphere would be a single object,
since only the surface was affected by the displacement map. Now, we
remove this restriction and consider the displacement map over arbitrary
points in 3D space. This allows the displacement map to generate
disconnected structures from the original sphere, which is necessary to
generate realistic-looking explosions.