Fractal Brownian Motion for realistic explosions
The displacement map is now defined in terms of Fractal Brownian Motion,
which is a noise function used to generate natural-looking noise
patterns. Fractal Brownian motion is a variant of fractal noise, which
is just noise (could be Perlin noise, Value noise, etc) rescaled and
added to itself. We use Value noise in our implementation. Overall,
this commit results in much more realistic-looking generated explosions.